Moved from helpers to sim and frame
